feat(2022): Simpler day 11
At the beginning of the original implementation, I wanted to parse
closures out of the strings to be able to run them directly.

That didn't work out; the code was colored through that view.

In this implementation, there are more types that handle their own
business, rather than being under the "Monkey" umbrella. :)